+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog
 =========
+## Unreleased
+### Fixed
+- Bug when attempting to resolve a transformer from a Yii config array
+
 ## 1.0.0-rc.4 - 2018-05-11
 ### Changed
 - Updated dependencies